+++ src/view/WelcomeView.java
@@ -1,30 +1,16 @@
 package view;
 
+import controller.createContent;
 import controller.AddPlayerListener;
 import controller.ExitListener;
-import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
-import javafx.geometry.HPos;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
-import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.GridPane;
-import javafx.scene.layout.StackPane;
-import javafx.scene.layout.VBox;
-import javafx.scene.paint.Color;
-import javafx.scene.paint.CycleMethod;
-import javafx.scene.paint.LinearGradient;
-import javafx.scene.paint.Stop;
-import javafx.scene.shape.Line;
-import javafx.scene.shape.Rectangle;
-import javafx.scene.text.Font;
-import javafx.scene.text.FontWeight;
-import javafx.scene.text.Text;
 import javafx.stage.Stage;
 
 import java.io.IOException;
@@ -32,8 +18,13 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
+/** NOTE: Perhaps Instead of constantly changing stages have one stage for the menu navigation
+ * have them all in one stage and switch scenes. Changing stages will resize the window
+ * regardless if they changed it previously. JavaFX has a section on scene switching.
+ */
+
 public class WelcomeView {
-	
+
 	Stage stage;
 	
 	public WelcomeView(Stage stage) {
@@ -50,7 +41,7 @@ public class WelcomeView {
 		root.setPadding(new Insets(25, 25, 25, 85));
 		root.setPrefSize(860, 600);
 		
-		//root.setGridLinesVisible(true);
+		root.setGridLinesVisible(true);
 
 		try (InputStream is = Files.newInputStream(Paths.get("assets/images/above-adventure-aerial-air.jpg"))) {
 			ImageView img = new ImageView(new Image(is));
@@ -60,25 +51,23 @@ public class WelcomeView {
 			System.out.println("Couldnt Load Image");
 		}
 
-
 		stage.setTitle("G2 Sabateur");
 
-
-		WelcomeView.Title title = new WelcomeView.Title("S A B A T E U R");
+		createContent.Title title = new createContent.Title("S A B A T E U R");
 		title.setId("title-text");
 
-		WelcomeView.MenuItem startGame = new WelcomeView.MenuItem("START GAME");
+		createContent.MenuItem startGame = new createContent.MenuItem("START GAME");
 		startGame.setOnMouseClicked(event -> {
 			System.out.println();
 		});
 
-		WelcomeView.MenuItem itemExit = new WelcomeView.MenuItem("EXIT");
+		createContent.MenuItem itemExit = new createContent.MenuItem("EXIT");
 		//itemExit.setOnMouseClicked(event -> System.exit(0));
 
-		WelcomeView.MenuBox vbox = new WelcomeView.MenuBox(
+		createContent.MenuBox vbox = new createContent.MenuBox(
 				startGame,
-				new WelcomeView.MenuItem("LOAD GAME"),
-				new WelcomeView.MenuItem("LEADERBOARD"),
+				new createContent.MenuItem("LOAD GAME"),
+				new createContent.MenuItem("LEADERBOARD"),
 				itemExit
 		);
 
@@ -105,83 +94,6 @@ public class WelcomeView {
 		stage.setScene(scene);
 		scene.getStylesheets().add(MainView.class.getResource("style.css").toExternalForm());
 				
-	}
-
-	private static class Title extends GridPane {
-		public Title(String name) {
-			// This is the main heading title box
-			Rectangle bg = new Rectangle(250, 60);
-			bg.setStroke(Color.WHITE);
-			bg.setStrokeWidth(2);
-			bg.setFill(null);
-
-			Text text = new Text(name);
-			text.setFill(Color.WHITE);
-			text.setFont(Font.font("Tw Cen MT Condensed", FontWeight.SEMI_BOLD, 50));
-
-			setHalignment(text, HPos.CENTER);
-			getChildren().addAll(bg, text);
-		}
-	}
-
-	private static class MenuBox extends VBox {
-		public MenuBox(WelcomeView.MenuItem... items) {
-			getChildren().add(createSeparator());
-
-			// Dynamically add separator after each menu item
-			for (WelcomeView.MenuItem item : items) {
-				getChildren().addAll(item, createSeparator());
-				setAlignment(Pos.CENTER);
-			}
-		}
-		private Line createSeparator() {
-			Line sep = new Line();
-			sep.setEndX(200);
-			sep.setStroke(Color.DARKGRAY);
-			return sep;
-		}
-	}
-
-	private static class MenuItem extends StackPane {
-		public MenuItem(String name) {
-			LinearGradient gradient = new LinearGradient(0, 0, 1, 0, true, CycleMethod.NO_CYCLE, new Stop[] {
-					new Stop(0, Color.DARKVIOLET),
-					new Stop(0.1, Color.BLACK),
-					new Stop(0.9, Color.BLACK),
-					new Stop(1, Color.DARKVIOLET)
-			});
-
-			Rectangle bg = new Rectangle(200, 30);
-			bg.setOpacity(0.4);
-
-			// Name that appears on the menu Item
-			Text text = new Text(name);
-			text.setFill(Color.DARKGRAY);
-			text.setFont(Font.font("Tw Cen MT Condensed", FontWeight.SEMI_BOLD, 22));
-
-			// Text is centered
-			setAlignment(Pos.CENTER);
-			getChildren().addAll(bg, text);
-
-			setOnMouseEntered(event -> {
-				bg.setFill(gradient);
-				text.setFill(Color.WHITE);
-			});
-
-			setOnMouseExited(event -> {
-				bg.setFill(Color.BLACK);
-				text.setFill(Color.DARKGRAY);
-			});
-
-			setOnMousePressed(event -> {
-				bg.setFill(Color.DARKVIOLET);
-			});
-
-			setOnMouseReleased(event -> {
-				bg.setFill(gradient);
-			});
-
-		}
 	}
 
 }
